Update README.md
There is a mistake that is easily made: copy and paste the three lines of shell code right after cloning the repo. You end up with the wrong path in the WON_CONFIG_DIR variable, leading to more problems. By adding the cd command, the shell code will work in both cases